requirements.txt
chromadb
sentence-transformers
fastapi
uvicorn
pillow
unstructured
python-docx
mammoth
pandas
openpyxl
xlrd
pypdf
ollama
